# [libc++] Optimize std::for_each_n for segmented iterators