This patch enhances the performance of `std::for_each_n` when used with segmented iterators, leading to significant performance improvements, summarized in the tables below. This addresses a subtask of https://github.com/llvm/llvm-project/issues/102817.

#### `std::deque` iterators

| Benchmark         | `deque<char>` | `deque<short>` | `deque<int>` |
|--------------------|-------------|--------------|------------|
| std::for_each_n   | 13.1x       | 17.7x        | 3.7x       |


#### `std::join_view` iterators

| Benchmark         | `vector<vector<char>>` | `vector<vector<short>>` | `vector<vector<int>>` |
|--------------------|-----------------------|-----------------------|---------------------|
| std::for_each_n   | 11.8x                | 13.9x                | 3.1x               |


 which results in performance improvements up to 17.7x for `std::deque<short>` iterators, and up to 13.9x for `join_view<vector<vector<short>>>` iterators. 



### Detailed Benchmarks

#### `std::deque` iterators
```
--------------------------------------------------------------------------
Benchmark                                    Before       After    Speedup
--------------------------------------------------------------------------
std::for_each_n(vector<char>)/8             4.26 ns     4.23 ns      1.0x
std::for_each_n(vector<char>)/32            2.68 ns     2.67 ns      1.0x
std::for_each_n(vector<char>)/50            9.49 ns     9.36 ns      1.0x
std::for_each_n(vector<char>)/1024          42.3 ns     40.1 ns      1.1x
std::for_each_n(vector<char>)/4096           163 ns      151 ns      1.1x
std::for_each_n(vector<char>)/8192           308 ns      294 ns      1.0x
std::for_each_n(vector<char>)/16384          608 ns      593 ns      1.0x
std::for_each_n(vector<char>)/65536         2435 ns     2464 ns      1.0x
std::for_each_n(vector<char>)/262144       10029 ns    10190 ns      1.0x
std::for_each_n(deque<char>)/8              6.57 ns     2.43 ns      2.7x
std::for_each_n(deque<char>)/32             24.0 ns     2.73 ns      8.8x
std::for_each_n(deque<char>)/50             33.2 ns     4.53 ns      7.3x
std::for_each_n(deque<char>)/1024            541 ns     44.9 ns     12.0x
std::for_each_n(deque<char>)/4096           2067 ns      169 ns     12.2x
std::for_each_n(deque<char>)/8192           4005 ns      305 ns     13.1x
std::for_each_n(deque<char>)/16384          7831 ns      639 ns     12.3x
std::for_each_n(deque<char>)/65536         31819 ns     2717 ns     11.7x
std::for_each_n(deque<char>)/262144       120801 ns    10674 ns     11.3x
std::for_each_n(list<char>)/8               4.97 ns     5.16 ns      1.0x
std::for_each_n(list<char>)/32              19.9 ns     20.6 ns      1.0x
std::for_each_n(list<char>)/50              40.6 ns     42.7 ns      1.0x
std::for_each_n(list<char>)/1024             996 ns     1038 ns      1.0x
std::for_each_n(list<char>)/4096            6186 ns     6341 ns      1.0x
std::for_each_n(list<char>)/8192           12522 ns    12391 ns      1.0x
std::for_each_n(list<char>)/16384          26158 ns    25739 ns      1.0x
std::for_each_n(list<char>)/65536         106410 ns   105299 ns      1.0x
std::for_each_n(list<char>)/262144        621473 ns   625741 ns      1.0x
std::for_each_n(vector<short>)/8            4.42 ns     3.92 ns      1.1x
std::for_each_n(vector<short>)/32           1.62 ns     1.64 ns      1.0x
std::for_each_n(vector<short>)/50           2.74 ns     2.75 ns      1.0x
std::for_each_n(vector<short>)/1024         34.0 ns     33.6 ns      1.0x
std::for_each_n(vector<short>)/4096          120 ns      117 ns      1.0x
std::for_each_n(vector<short>)/8192          229 ns      267 ns      0.9x
std::for_each_n(vector<short>)/16384         452 ns      469 ns      1.0x
std::for_each_n(vector<short>)/65536        2262 ns     2265 ns      1.0x
std::for_each_n(vector<short>)/262144       9129 ns     9140 ns      1.0x
std::for_each_n(deque<short>)/8             5.28 ns     1.78 ns      3.0x
std::for_each_n(deque<short>)/32            22.8 ns     2.08 ns     11.0x
std::for_each_n(deque<short>)/50            32.3 ns     4.46 ns      7.2x
std::for_each_n(deque<short>)/1024           545 ns     35.2 ns     15.5x
std::for_each_n(deque<short>)/4096          2158 ns      128 ns     16.9x
std::for_each_n(deque<short>)/8192          4303 ns      243 ns     17.7x
std::for_each_n(deque<short>)/16384         8624 ns      516 ns     16.7x
std::for_each_n(deque<short>)/65536        34569 ns     2336 ns     14.8x
std::for_each_n(deque<short>)/262144      137820 ns     9319 ns     14.8x
std::for_each_n(list<short>)/8              4.66 ns     4.95 ns      0.9x
std::for_each_n(list<short>)/32             19.9 ns     20.4 ns      1.0x
std::for_each_n(list<short>)/50             41.3 ns     41.1 ns      1.0x
std::for_each_n(list<short>)/1024           1018 ns     1021 ns      1.0x
std::for_each_n(list<short>)/4096           6110 ns     6294 ns      1.0x
std::for_each_n(list<short>)/8192          12433 ns    12692 ns      1.0x
std::for_each_n(list<short>)/16384         24739 ns    24820 ns      1.0x
std::for_each_n(list<short>)/65536        103376 ns   102812 ns      1.0x
std::for_each_n(list<short>)/262144       538314 ns   555664 ns      1.0x
std::for_each_n(vector<int>)/8              2.78 ns     2.73 ns      1.0x
std::for_each_n(vector<int>)/32             5.22 ns     5.26 ns      1.0x
std::for_each_n(vector<int>)/50             8.20 ns     8.65 ns      0.9x
std::for_each_n(vector<int>)/1024            156 ns      175 ns      0.9x
std::for_each_n(vector<int>)/4096            602 ns      758 ns      0.8x
std::for_each_n(vector<int>)/8192           1214 ns     1393 ns      0.9x
std::for_each_n(vector<int>)/16384          2417 ns     2690 ns      0.9x
std::for_each_n(vector<int>)/65536          9989 ns    10703 ns      0.9x
std::for_each_n(vector<int>)/262144        41512 ns    43798 ns      0.9x
std::for_each_n(deque<int>)/8               5.04 ns     2.75 ns      1.8x
std::for_each_n(deque<int>)/32              19.1 ns     5.56 ns      3.4x
std::for_each_n(deque<int>)/50              30.6 ns     8.55 ns      3.6x
std::for_each_n(deque<int>)/1024             567 ns      152 ns      3.7x
std::for_each_n(deque<int>)/4096            2241 ns      657 ns      3.4x
std::for_each_n(deque<int>)/8192            4512 ns     1334 ns      3.4x
std::for_each_n(deque<int>)/16384           9066 ns     2701 ns      3.4x
std::for_each_n(deque<int>)/65536          35955 ns    10887 ns      3.3x
std::for_each_n(deque<int>)/262144        146489 ns    44361 ns      3.3x
std::for_each_n(list<int>)/8                4.68 ns     6.05 ns      0.8x
std::for_each_n(list<int>)/32               21.0 ns     21.9 ns      1.0x
std::for_each_n(list<int>)/50               43.0 ns     42.2 ns      1.0x
std::for_each_n(list<int>)/1024             1015 ns     1035 ns      1.0x
std::for_each_n(list<int>)/4096             6373 ns     6331 ns      1.0x
std::for_each_n(list<int>)/8192            12757 ns    12836 ns      1.0x
std::for_each_n(list<int>)/16384           24879 ns    25035 ns      1.0x
std::for_each_n(list<int>)/65536          103931 ns   103773 ns      1.0x
std::for_each_n(list<int>)/262144         536841 ns   555330 ns      1.0x
--------------------------------------------------------------------------
```


#### `std::join_view` iterators
```
-----------------------------------------------------------------------------------------------------
Benchmark                                                        Before           After       Speedup
-----------------------------------------------------------------------------------------------------
std::for_each_n(join_view(vector<vector<char>>))/8              5.83 ns         2.56 ns          2.3x
std::for_each_n(join_view(vector<vector<char>>))/32             22.8 ns         3.12 ns          7.3x
std::for_each_n(join_view(vector<vector<char>>))/50             32.3 ns         5.37 ns          6.0x
std::for_each_n(join_view(vector<vector<char>>))/1024            477 ns         45.1 ns         10.6x
std::for_each_n(join_view(vector<vector<char>>))/4096           1898 ns          161 ns         11.8x
std::for_each_n(join_view(vector<vector<char>>))/8192           3785 ns          332 ns         11.4x
std::for_each_n(join_view(vector<vector<char>>))/16384          7530 ns          646 ns         11.7x
std::for_each_n(join_view(vector<vector<char>>))/65536         30685 ns         2670 ns         11.5x
std::for_each_n(join_view(vector<vector<char>>))/262144       122600 ns        10539 ns         11.6x
std::for_each_n(join_view(vector<vector<short>>))/8             6.14 ns         2.83 ns          2.2x
std::for_each_n(join_view(vector<vector<short>>))/32            25.2 ns         3.28 ns          7.7x
std::for_each_n(join_view(vector<vector<short>>))/50            33.7 ns         5.02 ns          6.7x
std::for_each_n(join_view(vector<vector<short>>))/1024           487 ns         38.6 ns         12.6x
std::for_each_n(join_view(vector<vector<short>>))/4096          1925 ns          150 ns         12.8x
std::for_each_n(join_view(vector<vector<short>>))/8192          3863 ns          291 ns         13.3x
std::for_each_n(join_view(vector<vector<short>>))/16384         7779 ns          558 ns         13.9x
std::for_each_n(join_view(vector<vector<short>>))/65536        30656 ns         2634 ns         11.6x
std::for_each_n(join_view(vector<vector<short>>))/262144      124937 ns        10667 ns         11.7x
std::for_each_n(join_view(vector<vector<int>>))/8               5.97 ns         3.45 ns          1.7x
std::for_each_n(join_view(vector<vector<int>>))/32              22.5 ns         7.20 ns          3.1x
std::for_each_n(join_view(vector<vector<int>>))/50              30.7 ns         10.7 ns          2.9x
std::for_each_n(join_view(vector<vector<int>>))/1024             491 ns          191 ns          2.6x
std::for_each_n(join_view(vector<vector<int>>))/4096            1928 ns          731 ns          2.6x
std::for_each_n(join_view(vector<vector<int>>))/8192            3874 ns         1402 ns          2.8x
std::for_each_n(join_view(vector<vector<int>>))/16384           7818 ns         2852 ns          2.7x
std::for_each_n(join_view(vector<vector<int>>))/65536          31101 ns        11439 ns          2.7x
std::for_each_n(join_view(vector<vector<int>>))/262144        126378 ns        45348 ns          2.8x
-----------------------------------------------------------------------------------------------------
```